requirements.txt
networkx==3.4.2
pandas==2.2.3
matplotlib==3.10.3
seaborn==0.13.2
numpy==1.26.4
pyyaml==6.0.2
python-louvain==0.16
powerlaw==1.5
tqdm==4.67.1
scipy==1.15.1
